Fix: only fix RBLNWorker
